rename migrate_config.py to populate_config.py
