+++ install_files/ansible-base/roles/build-securedrop-app-code-deb-pkg/files/usr.sbin.apache2
@@ -181,8 +181,8 @@
   /var/www/securedrop/journalist_templates/login.html r,
   /var/www/securedrop/journalist_templates/logo_upload_flashed.html r,
   /var/www/securedrop/journalist_templates/_source_row.html r,
-  /var/www/securedrop/migrate_config.py r,
-  /var/www/securedrop/migrate_config.pyc rw,
+  /var/www/securedrop/populate_config.py r,
+  /var/www/securedrop/populate_config.pyc rw,
   /var/www/securedrop/models.py r,
   /var/www/securedrop/models.pyc rw,
   /var/www/securedrop/request_that_secures_file_uploads.py r,
