+++ securedrop/tests/test_migrate_config.py
@@ -4,22 +4,22 @@ import pytest
 
 from os import path
 
-import migrate_config
+import populate_config
 
 
 def test_acquire_lock():
-    old_lock = migrate_config.LOCK_FILE
+    old_lock = populate_config.LOCK_FILE
 
     try:
-        migrate_config.LOCK_FILE = '/tmp/test-acquire-lock.lock'
+        populate_config.LOCK_FILE = '/tmp/test-acquire-lock.lock'
         # first attempt should succeed
-        with migrate_config.acquire_lock():
+        with populate_config.acquire_lock():
             # second attempt should fail
             with pytest.raises(SystemExit):
-                with migrate_config.acquire_lock():
+                with populate_config.acquire_lock():
                     pass
     finally:
-        migrate_config.LOCK_FILE = old_lock
+        populate_config.LOCK_FILE = old_lock
 
 
 def test_migrate_empty_case(tmpdir):
@@ -35,9 +35,9 @@ def test_migrate_empty_case(tmpdir):
 
     dummy = Dummy()
 
-    with mock.patch("migrate_config.import_config",
+    with mock.patch("populate_config.import_config",
                     return_value=dummy) as mock_import:
-        migrate_config.do_migration(True, config_dir)
+        populate_config.do_migration(True, config_dir)
 
     assert mock_import.called_once_with()
